--- a/sdk/mgmtcommon/ClientRuntime/ClientRuntime/Serialization/JsonConverterHelper.cs
+++ b/sdk/mgmtcommon/ClientRuntime/ClientRuntime/Serialization/JsonConverterHelper.cs
@@ -19,6 +19,8 @@ namespace Microsoft.Rest.Serialization
     /// </summary>
     public static class JsonConverterHelper
     {
+        private static readonly Regex splitCompiledRegex = new Regex(@"(?<!\\)\.", RegexOptions.Compiled);
+
         /// <summary>
         /// Serializes properties of the value object into JsonWriter.
         /// </summary>
@@ -108,21 +110,26 @@ namespace Microsoft.Rest.Serialization
                 throw new ArgumentNullException("property");
             }
 
-            string propertyName = property.PropertyName;
+            string propertyName = name;
             parentPath = new string[0];
 
             if (!string.IsNullOrEmpty(propertyName))
             {
-                string[] hierarchy = Regex.Split(propertyName, @"(?<!\\)\.")
-                    .Select(p => p?.Replace("\\.", ".")).ToArray();
+                string[] hierarchy = splitCompiledRegex.Split(propertyName);
+                for (int i = 0; i < hierarchy.Length; i++)
+                {
+                    hierarchy[i] = hierarchy[i]?.Replace("\\.", ".");
+                }
+
                 if (hierarchy.Length > 1)
                 {
-                    propertyName = hierarchy.Last();
-                    parentPath = hierarchy.Take(hierarchy.Length - 1).ToArray();
+                    propertyName = hierarchy[hierarchy.Length - 1];
+                    Array.Resize(ref hierarchy, hierarchy.Length - 1);
+                    parentPath = hierarchy;
                 }
             }
 
-            return propertyName;            
+            return propertyName;       
         }
 
         public static bool IsJsonExtensionData(this JsonProperty property)